demo/requirements.txt
Django == 1.9.7
pymysql == 0.7.4
django-jsonfield
djangorestframework == 3.3.3
djangorestframework-jwt
django-admin-bootstrapped
django-enumfields == 0.7.4
requests == 2.10.0
drf-nested-routers
django-celery
dynamic_rest
drf-enum-field
django-model-utils
qiniu
django-guardian
django-debug-toolbar
celery==3.1.18
pycrypto

 
